feat(budget): 사업별 편성률 2버킷을 소수까지 받는다 (MIG-15)
ASG_RT는 소수 5자리를 담고 비목별 경로(ioeRates)는 이미 BigDecimal이지만,
2버킷(assetDupRt·costDupRt)만 Integer였다. Jackson이 70.5를 조용히 70으로
잘라 /budget/work 화면에서 소수 편성률을 넣을 방법이 없었다.

두 필드를 BigDecimal로 넓히고 BudgetRateApplicationService의
BigDecimal.valueOf 감싸기를 제거한다. 프론트 입력 컨트롤 대응은 it_frontend
커밋에서 이어진다.

+++ src/main/java/com/kdb/it/domain/budget/work/service/BudgetRateApplicationService.java
@@ -157,14 +157,8 @@ public class BudgetRateApplicationService {
         int nextSno = 0;
         int totalRecords = 0;
         for (BudgetWorkDto.ItemRate item : request.items()) {
-            BigDecimal assetRate =
-                    item.assetDupRt() != null
-                            ? BigDecimal.valueOf(item.assetDupRt())
-                            : DEFAULT_DUP_RT;
-            BigDecimal costRate =
-                    item.costDupRt() != null
-                            ? BigDecimal.valueOf(item.costDupRt())
-                            : DEFAULT_DUP_RT;
+            BigDecimal assetRate = item.assetDupRt() != null ? item.assetDupRt() : DEFAULT_DUP_RT;
+            BigDecimal costRate = item.costDupRt() != null ? item.costDupRt() : DEFAULT_DUP_RT;
             Map<String, BigDecimal> ioeRates = item.ioeRates() == null ? Map.of() : item.ioeRates();
             if ("BPROJM".equals(item.orcTb())) {
                 for (Bitemm source :
